chore(FE): rename snake references to spider
